settings.py
import os
from dotenv import load_dotenv

load_dotenv()

START_GG_API_KEY = os.environ.get("START_GG_API_KEY")
START_GG_API_URL = os.environ.get("START_GG_API_URL")
REDDIT_CLIENT_ID = os.environ.get("REDDIT_CLIENT_ID")
REDDIT_CLIENT_SECRET = os.environ.get("REDDIT_CLIENT_SECRET")
REDDIT_USERNAME = os.environ.get("REDDIT_USERNAME")
REDDIT_PASSWORD = os.environ.get("REDDIT_PASSWORD")
REDIS_HOST = os.environ.get("REDIS_HOST")
REDIS_PORT = os.environ.get("REDIS_PORT")
