requirements.txt
requests
python-dotenv
pytest
inflect
praw
redis
pytz
coverage
